feat: Added date format validation
The regex used in this code was referenced from https://www.w3resource.com/javascript/form/javascript-date-validation.php but was modified to suit the needs of the travel app input. Originally this regex checked for a dd/mm/yyyy format but will now look for a mm/dd/yyyy format.